% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combForecast.R
\name{combForecast}
\alias{combForecast}
\title{Forecast reconciliation}
\usage{
combForecast(forecast, residual, weights = c("wls", "cwls", "bshr", "pshr",
  "mint"), linearComb = S)
}
\arguments{
\item{forecast}{A matrix with a dimension of p (row) and h (column), where p is the number of time series and h is the forecast horizon}

\item{residual}{A matrix with a dimension of p (row) and T (column), where p is the number of time series and T is the number of observations}

\item{weights}{A reconciliation weight matrix.}

\item{linearComb}{A matrix of a linear combination. In hierarchical forecasting, this is the summation matrix or S.}
}
\value{
Function returns the following variables:
\itemize{
\item{\code{reconciledForecast} - the reconciled forecasts}
\item{\code{baseForecast} - the base forecasts}
\item{\code{reconMatrix} - the reconciliation weights matrix}
\item{\code{smatrix} - the linear combination. Usually, it is the summation matrix}
\item{\code{invcovMatrix} - the inverse of the covariance matrix estimation approximation}
}
}
\description{
Forecast reconciliation
}
