% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lambda.R
\name{estimate_lambda}
\alias{estimate_lambda}
\title{Estimate a hyper-parameter in the shrinkage estimators for exponential smoothing models}
\usage{
estimate_lambda(object, x0 = 0.1, loss = c("RIDGE", "LASSO"),
  origins = 5)
}
\arguments{
\item{object}{An adam object}

\item{x0}{An initialisation for lambda. Should be equal or more than 0 and less than 1}

\item{loss}{A string of the loss function, either "RIDGE" or "LASSO". The default is "RIDGE"}

\item{origins}{A number of forecast origins. The default is the frequency of the time series}
}
\value{
Function returns the following variables:
\itemize{
\item{\code{lambda_{min}} - the optimal hyper-parameter of ETS}
\item{\code{model} - the model}
\item{\code{loss} - the loss function, either "RIDGE" or "LASSO"}
\item{\code{data} - the training set}
}
}
\description{
Estimate a hyper-parameter in the shrinkage estimators for exponential smoothing models
}
\examples{

library(smooth)
library(greybox)

# Estimate an exponential smoothing model first
fit <- adam(AirPassengers, loss = "MSE")

# Implement estimate_lambda() to find the optimal hyper-parameter
choosingLambda <- estimate_lambda(fit, x0 = 0.5, loss = "RIDGE", origins = 5)

# Inspect the result
choosingLambda

# Re-estimate adam() with the result of estimate_lambda()
fitShrinkage <- adam(choosingLambda$data, model = choosingLambda$model,
                     loss = choosingLambda$loss,
                     lambda = choosingLambda$lambda_min)

# Inspect the effect of smoothing parameter shrinkage
fit$persistence
fitShrinkage$persistence

}
\references{
\itemize{
\item Pritularga, K., Svetunkov, I., Kourentzes, N., (2022) Shrinkage estimator for exponential smoothing models. International Journal of Forecasting, XX, pp. XX.
}
}
\author{
Kandrika Pritularga
}
